refactor/#510: date 계산 함수 util 분리
