+++ components/statistics/BlogStatistic.tsx
@@ -2,20 +2,13 @@
 import React from 'react';
 import styled from 'styled-components';
 
+import { getCurrentDateTime } from '@/utils/getCurrentDateTime';
+
 import VisitantUI from './BlogStatistics/VisitantUI';
 
 const BlogStatistic = () => {
-  const getCurrentDate = (): string => {
-    const today = new Date();
-    const year = today.getFullYear();
-    const month = String(today.getMonth() + 1).padStart(2, '0');
-    const day = String(today.getDate()).padStart(2, '0');
-    const hours = String(today.getHours()).padStart(2, '0');
-    const minutes = String(today.getMinutes()).padStart(2, '0');
-    return `${year}.${month}.${day} ${hours}:${minutes}`;
-  };
+  const currentDate = getCurrentDateTime();
 
-  const currentDate = getCurrentDate();
   return (
     <div>
       <TitleContainer>
